requirements.txt
python-dotenv==1.0.0
openai==1.65.4
streamlit==1.32.0
matplotlib==3.8.0
sympy==1.12
pygments==2.16.1 